+++ app/src/main/java/org/sopt/zooczoocbbangbbang/presentation/main/record/MissionFragment.kt
@@ -3,9 +3,7 @@ package org.sopt.zooczoocbbangbbang.presentation.main.record
 import android.os.Bundle
 import android.util.Log
 import android.view.View
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.viewModels
-import coil.load
 import org.sopt.zooczoocbbangbbang.R
 import org.sopt.zooczoocbbangbbang.databinding.FragmentMissionBinding
 import org.sopt.zooczoocbbangbbang.presentation.base.BindingFragment
@@ -16,22 +14,8 @@ class MissionFragment : BindingFragment<FragmentMissionBinding>(R.layout.fragmen
         super.onViewCreated(view, savedInstanceState)
         binding.lifecycleOwner = viewLifecycleOwner
         binding.viewmodel = missionViewModel
-        clickImageBtn()
         clickNextBtn()
         observe()
-    }
-
-    // 사진을 불러오는 함수
-    private fun clickImageBtn() {
-        val launcher = registerForActivityResult(ActivityResultContracts.GetContent()) {
-            binding.ivMissionBgimg.load(it) {
-                transformations()
-                missionViewModel.image.value = it
-            }
-        }
-        binding.ivMissionImgbtn.setOnClickListener {
-            launcher.launch("image/*")
-        }
     }
 
     private fun clickNextBtn() {
