+++ app/src/main/res/layout/fragment_mission_view_pager.xml
@@ -1,0 +1,108 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context=".presentation.main.record.MissionViewPagerFragment">
+    <data>
+        <import type="androidx.core.content.ContextCompat" />
+
+        <variable
+            name="viewmodel"
+            type="org.sopt.zooczoocbbangbbang.presentation.main.record.MissionViewModel" />
+    </data>
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/cl_mission_basebg"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:id="@+id/cl_mission_bgwhite"
+            android:layout_width="0dp"
+            android:layout_height="0dp"
+            android:layout_marginHorizontal="30dp"
+            android:background="@drawable/shape_record_white_radius_24"
+            android:backgroundTint="@color/white"
+            app:layout_constraintDimensionRatio="300:460"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent">
+
+            <TextView
+                android:id="@+id/tv_mission_question"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="@string/mission_question"
+                android:textColor="@color/grey_4f"
+                android:layout_marginHorizontal="45dp"
+                android:gravity="center"
+                style="@style/Typography.M16"
+                android:layout_marginTop="24dp"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent" />
+
+            <androidx.cardview.widget.CardView
+                android:id="@+id/cv_mission_photo"
+                android:layout_width="0dp"
+                android:layout_height="0dp"
+                android:layout_marginHorizontal="18dp"
+                android:backgroundTint="#F5F5F5"
+                app:cardCornerRadius="20dp"
+                app:cardElevation="0dp"
+                android:layout_marginTop="16dp"
+                app:layout_constraintDimensionRatio="264:230"
+                app:layout_constraintEnd_toEndOf="@+id/cl_mission_bgwhite"
+                app:layout_constraintStart_toStartOf="@+id/cl_mission_bgwhite"
+                app:layout_constraintTop_toBottomOf="@+id/tv_mission_question">
+
+                <ImageView
+                    android:id="@+id/iv_mission_imgbtn"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:layout_marginVertical="54dp"
+                    android:layout_marginStart="91dp"
+                    android:layout_marginEnd="102dp"
+                    android:scaleType="fitCenter"
+                    app:srcCompat="@drawable/img_record_photo_24" />
+
+                <ImageView
+                    android:id="@+id/iv_mission_bgimg"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:scaleType="fitXY" />
+
+            </androidx.cardview.widget.CardView>
+
+            <EditText
+                android:id="@+id/edt_mission"
+                android:layout_width="0dp"
+                android:layout_height="0dp"
+                app:layout_constraintDimensionRatio="264:112"
+                android:layout_marginHorizontal="18dp"
+                android:layout_marginTop="12dp"
+                android:layout_marginBottom="18dp"
+                android:text=""
+                android:background="@drawable/shape_rec_white2_radius_12"
+                android:hint="@string/mission_hintText"
+                android:inputType="textMultiLine"
+                android:paddingHorizontal="18dp"
+                android:paddingTop="18dp"
+                android:paddingBottom="18dp"
+                android:gravity="top"
+                android:textAppearance="@style/Typography.M14"
+                android:textColorHint="#9C9898"
+                android:lineSpacingExtra="5sp"
+                tools:ignore="TouchTargetSizeCheck"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                android:layout_marginStart="18dp"
+                android:layout_marginEnd="18dp"
+                app:layout_constraintTop_toBottomOf="@+id/cv_mission_photo"
+                app:layout_constraintBottom_toBottomOf="parent" />
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
+
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+</layout>
