+++ app/src/main/res/values/strings.xml
@@ -21,5 +21,7 @@
     <!--미션하기 뷰-->
     <string name="mission_question">갤러리를 볼 때마다 당신을 웃게 \n만드는 강아지 사진은 무엇인가요?</string>
     <string name="mission_hintText">ex) 2023년 2월 30일 \n어떤 순간이었는지 가족에게 남겨주세요</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 
 </resources>
